+++ app/src/main/AndroidManifest.xml
@@ -23,6 +23,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".DeviceActivity2" />
+        <activity android:name=".DeviceActivity" />
     </application>
 </manifest>
